Backend/requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Werkzeug==2.3.7

# Database
pymongo==4.5.0

# Authentication and security
PyJWT==2.8.0
bcrypt==4.0.1

# Machine Learning and Data Science
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# Data visualization (optional)
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Environment and configuration
python-dotenv==1.0.0

# Development and testing
pytest==7.4.0
pytest-flask==1.2.0

# Logging and monitoring
gunicorn==21.2.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2